get currencyRequest from frontend + slight refactoring
