+++ src/main/java/com/example/demo/service/APICall.java
@@ -1,4 +1,4 @@
-package com.example.demo.entity;
+package com.example.demo.service;
 
 public enum APICall {
     GET_CARDS,
