+++ src/main/java/com/example/demo/service/APIService.java
@@ -1,7 +1,9 @@
 package com.example.demo.service;
 
-import com.example.demo.entity.PokemonCard;
-import com.example.demo.entity.PokemonCardDeck;
+import com.example.demo.dto.CurrencyRequest;
+import com.example.demo.dto.PokemonCardRequest;
+import com.example.demo.dto.PokemonDeckRequest;
+import com.example.demo.dto.PokemonDeckResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -17,15 +19,19 @@ public class APIService {
     private final ProductService productService;
     private final CurrencyService currencyService;
 
-    public List<PokemonCard> getPokemonCardList() {
-        return productService.getPokemonCardList();
+    public List<PokemonCardRequest> getPokemonCards() {
+        return productService.getPokemonCards();
     }
 
-    public List<PokemonCardDeck> getPokemonCardDeckList() {
-        return productService.pokemonCardDeckList();
+    public List<PokemonDeckRequest> getPokemonDecks() {
+        return productService.getPokemonDecks();
     }
 
-    public BigDecimal getCurrency(String currency, String price) {
-        return currencyService.getCurrency(currency, price);
+    public BigDecimal getPriceInCurrency(CurrencyRequest currencyRequest) {
+        return currencyService.getPriceInCurrency(currencyRequest);
+    }
+
+    public PokemonDeckResponse createPokemonDeck(String pokemonCardList) {
+        return productService.createPokemonDeck(pokemonCardList);
     }
 }
